v3.0

+++ starter/src/compute/compute_init.sh
@@ -19,66 +19,14 @@ echo "ARCH=$ARCH"
 sudo setenforce 0
 sudo sed -i s/^SELINUX=.*$/SELINUX=permissive/ /etc/selinux/config
 
-# GrowFS - resize the volume
+# Resize the boot volume (if >47GB)
 sudo /usr/libexec/oci-growfs -y
 
 # Set VI and NANO in utf8
 echo "export LC_CTYPE=en_US.UTF-8" >> $HOME/.bashrc
 
-# -- Shared Install function ------------------------------------------------
-
-install_java() {
-  # Install the JVM (jdk or graalvm)
-  if [ "$TF_VAR_java_vm" == "graalvm" ]; then
-    # GraalVM
-    if [ "$TF_VAR_java_version" == 8 ]; then
-      sudo dnf install -y graalvm21-ee-8-jdk 
-      sudo update-alternatives --set java /usr/lib64/graalvm/graalvm22-ee-java8/bin/java
-    elif [ "$TF_VAR_java_version" == 11 ]; then
-      sudo dnf install -y graalvm22-ee-11-jdk
-      sudo update-alternatives --set java /usr/lib64/graalvm/graalvm22-ee-java11/bin/java
-    elif [ "$TF_VAR_java_version" == 17 ]; then
-      sudo dnf install -y graalvm22-ee-17-jdk 
-      sudo update-alternatives --set java /usr/lib64/graalvm/graalvm22-ee-java17/bin/java
-      # sudo update-alternatives --set native-image /usr/lib64/graalvm/graalvm22-ee-java17/lib/svm/bin/native-image
-    elif [ "$TF_VAR_java_version" == 21 ]; then
-      sudo dnf install -y graalvm-21-jdk
-      sudo update-alternatives --set java /usr/lib64/graalvm/graalvm-java21/bin/java
-      # sudo update-alternatives --set native-image /usr/lib64/graalvm/graalvm-java21/lib/svm/bin/native-image
-    else
-      sudo dnf install -y graalvm-25-jdk
-      sudo update-alternatives --set java /usr/lib64/graalvm/graalvm-java25/bin/java
-      # sudo update-alternatives --set native-image /usr/lib64/graalvm/graalvm-java21/lib/svm/bin/native-image
-    fi   
-  else
-    # JDK 
-    # Needed due to concurrency
-    sudo dnf install -y alsa-lib 
-    if [ "$TF_VAR_java_version" == 8 ]; then
-      sudo dnf install -y java-1.8.0-openjdk
-    elif [ "$TF_VAR_java_version" == 11 ]; then
-      sudo dnf install -y java-11  
-    elif [ "$TF_VAR_java_version" == 17 ]; then
-      sudo dnf install -y java-17        
-    elif [ "$TF_VAR_java_version" == 21 ]; then
-      sudo dnf install -y java-21         
-    else
-      sudo dnf install -y java-25  
-      # Trick to find the path
-      # cd -P "/usr/java/latest"
-      # export JAVA_LATEST_PATH=`pwd`
-      # cd -
-      # sudo update-alternatives --set java $JAVA_LATEST_PATH/bin/java
-    fi
-  fi
-
-  # JMS agent deploy (to fleet_ocid )
-  if [ -f jms_agent_deploy.sh ]; then
-    chmod +x jms_agent_deploy.sh
-    sudo ./jms_agent_deploy.sh
-  fi
-}
-export -f install_java
+# Shared Install Funciton
+. ./install_util.sh
 
 # -- App --------------------------------------------------------------------
 # Application Specific installation
@@ -89,32 +37,35 @@ for APP_DIR in `ls -d app* | sort -g`; do
   if [ -f $APP_DIR/install.sh ]; then
     echo "-- $APP_DIR: Install -----------------------------------------"
     chmod +x ${APP_DIR}/install.sh
+    chmod +x ${APP_DIR}/env.sh
     ${APP_DIR}/install.sh
   fi  
 done
 
 # -- app/start*.sh -----------------------------------------------------------
 for APP_DIR in `ls -d app* | sort -g`; do
-  echo "#!/usr/bin/env bash" > $APP_DIR/restart.sh 
-  chmod +x $APP_DIR/restart.sh  
-  for START_SH in `ls $APP_DIR/start*.sh | sort -g`; do
-    echo "-- $START_SH ---------------------------------------"
-    if [[ "$START_SH" =~ start_(.*).sh ]]; then
-      APP_NAME=$(echo "$START_SH" | sed -E 's/(.*)\/start_([a-zA-Z0-9_]+)\.sh$/\1_\2/')
-    else
-      APP_NAME=${APP_DIR}
-    fi
-    echo "APP_NAME=$APP_NAME"
-    # Hardcode the connection to the DB in the start.sh
-    if [ "$DB_URL" != "" ]; then
-      sed -i "s!##JDBC_URL##!$JDBC_URL!" $START_SH 
-      sed -i "s!##DB_URL##!$DB_URL!" $START_SH 
-    fi  
-    sed -i "s!##TF_VAR_java_vm##!$TF_VAR_java_vm!" $START_SH
-    chmod +x $START_SH
+  if [ -f $APP_DIR/restart.sh ]; then
+    echo "$APP_DIR/restart.sh exists already"
+  else
+    echo "#!/usr/bin/env bash" > $APP_DIR/restart.sh 
+    for START_SH in `ls $APP_DIR/start*.sh | sort -g`; do
+      echo "-- $START_SH ---------------------------------------"
+      if [[ "$START_SH" =~ start_(.*).sh ]]; then
+        APP_NAME=$(echo "$START_SH" | sed -E 's/(.*)\/start_([a-zA-Z0-9_]+)\.sh$/\1_\2/')
+      else
+        APP_NAME=${APP_DIR}
+      fi
+      echo "APP_NAME=$APP_NAME"
+      # Hardcode the connection to the DB in the start.sh
+      if [ "$DB_URL" != "" ]; then
+        sed -i "s!##JDBC_URL##!$JDBC_URL!" $START_SH 
+        sed -i "s!##DB_URL##!$DB_URL!" $START_SH 
+      fi  
+      sed -i "s!##TF_VAR_java_vm##!$TF_VAR_java_vm!" $START_SH
+      chmod +x $START_SH
 
-    # Create an "app.service" that starts when the machine starts.
-     cat > /tmp/$APP_NAME.service << EOT
+      # Create an "app.service" that starts when the machine starts.
+      cat > /tmp/$APP_NAME.service << EOT
 [Unit]
 Description=App
 After=network.target
@@ -128,13 +79,15 @@ User=opc
 [Install]
 WantedBy=default.target
 EOT
-    sudo cp /tmp/$APP_NAME.service /etc/systemd/system
-    sudo chmod 664 /etc/systemd/system/$APP_NAME.service
-    sudo systemctl daemon-reload
-    sudo systemctl enable $APP_NAME.service
-    sudo systemctl restart $APP_NAME.service
-    echo "sudo systemctl restart $APP_NAME" >> $APP_DIR/restart.sh 
-  done  
+      sudo cp /tmp/$APP_NAME.service /etc/systemd/system
+      sudo chmod 664 /etc/systemd/system/$APP_NAME.service
+      sudo systemctl daemon-reload
+      sudo systemctl enable $APP_NAME.service
+      echo "sudo systemctl restart $APP_NAME" >> $APP_DIR/restart.sh 
+    done  
+  fi  
+  chmod +x $APP_DIR/restart.sh  
+  $APP_DIR/restart.sh
 done 
 
 # -- Helper --------------------------------------------------------------------
